requirements.txt

fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
qdrant-client
langchain
openai
paddleocr
paddlepaddle
pdfplumber
python-docx
PyMuPDF 
python-dotenv


streamlit


requests
aiofiles
httpx
scikit-learn
numpy
tqdm
Pillow

# Optional: To avoid future dependency resolution issues
pip>=23.0
setuptools>=65.0
